+++ lib/graphql/static_validation/rules/required_arguments_are_present.rb
@@ -21,7 +21,7 @@ module GraphQL
         present_argument_names = ast_node.arguments.map(&:name)
         required_argument_names = context.warden.arguments(defn)
           .select { |a| a.type.kind.non_null? && !a.default_value? && context.warden.get_argument(defn, a.name) }
-          .map(&:name)
+          .map!(&:name)
 
         missing_names = required_argument_names - present_argument_names
         if missing_names.any?