+++ spec/integration/rails/graphql/schema_spec.rb
@@ -149,12 +149,13 @@ describe GraphQL::Schema do
       it "contains built-in directives" do
         schema = GraphQL::Schema
 
-        assert_equal ['deprecated', 'include', 'oneOf', 'skip'], schema.directives.keys.sort
+        assert_equal ['deprecated', 'include', 'oneOf', 'skip', 'specifiedBy'], schema.directives.keys.sort
 
         assert_equal GraphQL::Schema::Directive::Deprecated, schema.directives['deprecated']
         assert_equal GraphQL::Schema::Directive::Include, schema.directives['include']
         assert_equal GraphQL::Schema::Directive::Skip, schema.directives['skip']
         assert_equal GraphQL::Schema::Directive::OneOf, schema.directives['oneOf']
+        assert_equal GraphQL::Schema::Directive::SpecifiedBy, schema.directives['specifiedBy']
       end
     end
   end