+++ spec/spec_helper.rb
@@ -32,7 +32,7 @@ end
 Minitest::Spec.make_my_diffs_pretty!
 
 module CheckWardenShape
-  DEFAULT_SHAPE = GraphQL::Schema::Warden.new(nil, context: {}, schema: GraphQL::Schema).instance_variables
+  DEFAULT_SHAPE = GraphQL::Schema::Warden.new(context: {}, schema: GraphQL::Schema).instance_variables
 
   class CheckShape
     def initialize(warden)